requirements.txt
Django==4.2.1
djangorestframework==3.14.0
django-rest-knox==4.2.0
dj-database-url==1.2.0
drf-yasg==1.21.4
black
isort
flake8